Azure ve Filter

+++ Infrastructure/Repositories/ProductService.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using backend.Application.DTOs.Filter;
 using backend.Application.DTOs.Product;
 using backend.Application.Services;
 using backend.Domain.Entities;
@@ -15,28 +16,64 @@ namespace backend.Infrastructure.Repositories
     {
         private readonly IMapper _mapper;
         private readonly IService<Product> _service;
-        public ProductService(IMapper mapper, IService<Product> service = null)
+        private readonly IBlobService _blob;
+
+        public ProductService(IMapper mapper, IService<Product> service, IBlobService blob)
         {
             _mapper = mapper;
             _service = service;
+            _blob = blob;
         }
 
-        public async Task<Response<NoContent>> AddProductAsync(CreateProductRequestDto product)
+        public async Task<Response<NoContent>> AddProductAsync(CreateProductRequestDto dto)
         {
-            var newProduct = _mapper.Map<CreateProductRequestDto, Product>(product);
-            await _service.AddAsync(newProduct);
-            return Response<NoContent>.Success(HttpStatusCode.OK, "Ürün Ekleme Başarılı!");
+            if (dto.Image == null || dto.Image.Length == 0) { 
+                //return Response<NoContent>.Fail("Resim zorunlu.", HttpStatusCode.BadRequest);
+                var productt = _mapper.Map<Product>(dto);
+                productt.ImageUrl = "BU KISIM TEST İÇİN SİLİNECEK";
+
+                await _service.AddAsync(productt);
+                return Response<NoContent>.Success(HttpStatusCode.Created, "Ürün eklendi");
+            }
+            string url = await _blob.UploadAsync(dto.Image);
+
+            var product = _mapper.Map<Product>(dto);
+            product.ImageUrl = url;
+
+            await _service.AddAsync(product);
+            return Response<NoContent>.Success(HttpStatusCode.Created, "Ürün eklendi");
         }
 
         public async Task<Response<NoContent>> DeleteProductAsync(int productId)
         {
+            var product = await _service.GetByIdAsync(productId);
+            if (product == null)
+                return Response<NoContent>.Fail("Ürün bulunamadı", HttpStatusCode.NotFound);
+
+            if (!string.IsNullOrEmpty(product.ImageUrl))
+                await _blob.DeleteAsync(Path.GetFileName(product.ImageUrl));
+
             await _service.DeleteAsync(productId);
-            return Response<NoContent>.Success(HttpStatusCode.OK, "Silme Başarılı!");
+            return Response<NoContent>.Success(HttpStatusCode.OK, "Silindi");
         }
 
-        public async Task<Response<IEnumerable<GetProductResponseDto>>> GetProductAllAsync()
+        public async Task<Response<IEnumerable<GetProductResponseDto>>> GetProductAllAsync(ProductFilter filter)
         {
-            var products = await _service.GetAllAsync();
+            var query = _service.Query().AsQueryable();
+
+            if (filter.CategoryId != null)
+                query = query.Where(p => p.CategoryId == filter.CategoryId);
+
+            if (!string.IsNullOrEmpty(filter.Name))
+                query = query.Where(p => p.Name.ToLower().Contains(filter.Name.ToLower()));
+
+            if (filter.MinPrice != null)
+                query = query.Where(p => p.Price >= filter.MinPrice);
+
+            if (filter.MaxPrice != null)
+                query = query.Where(p => p.Price <= filter.MaxPrice);
+
+            var products = await query.ToListAsync();
             var response = _mapper.Map<IEnumerable<Product>, IEnumerable<GetProductResponseDto>>(products);
             return Response<IEnumerable<GetProductResponseDto>>.Success(response, HttpStatusCode.OK, "Başarılı!");
         }
@@ -55,16 +92,24 @@ namespace backend.Infrastructure.Repositories
             }
         }
 
-        public async Task<Response<NoContent>> UpdateProductAsync(UpdateProductRequestDto newProduct)
+        public async Task<Response<NoContent>> UpdateProductAsync(UpdateProductRequestDto dto)
         {
-            var product = await _service.GetByIdAsync(newProduct.Id);
+            var product = await _service.GetByIdAsync(dto.Id);
             if (product == null)
+                return Response<NoContent>.Fail("Ürün bulunamadı", HttpStatusCode.NotFound);
+
+            _mapper.Map(dto, product);
+
+            if (dto.Image is not null && dto.Image.Length > 0)
             {
-                return Response<NoContent>.Fail("Güncellenecek ürün bulunamadı.", HttpStatusCode.BadRequest);
+                if (!string.IsNullOrEmpty(product.ImageUrl))
+                    await _blob.DeleteAsync(Path.GetFileName(product.ImageUrl));
+
+                product.ImageUrl = await _blob.UploadAsync(dto.Image);
             }
-            _mapper.Map(newProduct, product); 
-            await _service.UpdateAsync(product); 
-            return Response<NoContent>.Success(HttpStatusCode.OK, "Güncelleme başarılı!");
+
+            await _service.UpdateAsync(product);
+            return Response<NoContent>.Success(HttpStatusCode.OK, "Güncellendi");
         }
 
         public async Task<bool> IsProductNameExist(string name)
@@ -72,6 +117,34 @@ namespace backend.Infrastructure.Repositories
             if (await _service.GetFirstOrDefaultAsync(p => p.Name == name) != null)
                 return true;
             return false;
+        }
+
+        public async Task<Response<IEnumerable<GetProductResponseDto>>> GetProductByCategoryAsync(int categoryId)
+        {
+            var products = await _service.Query().Where(p => p.CategoryId == categoryId).OrderBy(p => p.Name).ToListAsync();
+            if (products == null)
+            {
+                return Response<IEnumerable<GetProductResponseDto>>.Fail("Ürün Yok.", HttpStatusCode.BadRequest);
+            }
+            else
+            {
+                var response = _mapper.Map<IEnumerable<Product>, IEnumerable<GetProductResponseDto>>(products);
+                return Response<IEnumerable<GetProductResponseDto>>.Success(response, HttpStatusCode.OK, "Başarılı!");
+            }
+        }
+
+        public async Task<Response<IEnumerable<GetProductResponseDto>>> GetProductByNameAsync(string name)
+        {
+            var products = await _service.Query().Where(p => p.Name.ToLower().Contains(name.ToLower())).OrderBy(p => p.Name).ToListAsync();
+            var response = _mapper.Map<IEnumerable<Product>, IEnumerable<GetProductResponseDto>>(products);
+            return Response<IEnumerable<GetProductResponseDto>>.Success(response, HttpStatusCode.OK, "Başarılı!");
+        }
+
+        public async Task<Response<IEnumerable<GetProductResponseDto>>> GetProductByPriceRange(int minPrice, int maxPrice)
+        {
+            var products = await _service.Query().Where(p => p.Price>=minPrice && p.Price <= maxPrice).OrderBy(p => p.Price).ToListAsync();
+            var response = _mapper.Map<IEnumerable<Product>, IEnumerable<GetProductResponseDto>>(products);
+            return Response<IEnumerable<GetProductResponseDto>>.Success(response, HttpStatusCode.OK, "Başarılı!");
         }
     }
 }
